Rename to ProjectFiltersModal
